+++ DataAccessTier/Repository/JobManagement/ClusterProjectRepository.cs
@@ -37,6 +37,8 @@ internal class ClusterProjectRepository : GenericRepository<ClusterProject>, ICl
     {
         return await _context.ClusterProjects
             .Include(x => x.ClusterProjectCredentials)
+            .Include(x => x.Cluster)
+            .Include(x => x.Project)
             .Where(cp => cp.ProjectId == projectId && cp.ClusterId == clusterId)
             .FirstOrDefaultAsync();
     }